1. Bug Fixing Itinerary Save dan Delete 2. Add Transportation Service 3. Add Riview Service 4. Bug Fixing add Id

+++ src/main/java/liburind/project/model/Riview.java
@@ -13,7 +13,7 @@ public class Riview {
 	
 	@Id
 	private String riviewId;
-	private Integer riviewScore;
+	private float riviewScore;
 	private String riviewDetail;
 	private LocalDateTime riviewRecordedTime;
 	private String userId;
